refactor(test): migrate storage helpers to three-tier test architecture
Move DB-direct write functions to db-test-seeders/storage.ts and read-only
query functions to db-test-assertions/storage.ts. Keep API-based helpers
(createTestArtifact, createTestVolume, createTestMemory) in api-test-helpers
with re-exports for backward compatibility.

Closes #9354
